requirements.txt
Django==5.0.6
Pillow==10.4.0
python-dotenv==1.0.1
rembg==2.0.51
pandas==2.2.2
celery==5.4.0
redis==5.0.4
gunicorn==22.0.0